feat: Adding the database handling functions
Creating a base database handling. The main features implemented are:
- Database initialization
- Adding new items to the table
- Updating items from the table by id
- Searching items by id
- Searching items by status
- Listing all the items
- Deleting items from table by id

Future improvements to make to the model is to create a class
where the function will be used as methods, this would result
in a better code structure, reducing the redundant code,
for example the constant need of creating and closing the connection
to the database, the same for the cursor.